v0.1.3 — 依赖升级至最新兼容版本
移除所有 = 版本锁定，使用 semver 范围：
- tonic 0.9→0.12, prost 0.11→0.13, tonic-build 0.9→0.12
- tokio 1.36→1.x (最新), 移除未使用的 tokio-stream
- warp 0.3→0.4 (.run→.bind 适配)
- reqwest 0.11→0.12
- base64 0.21→0.22
- dashmap 5→6
- thiserror 1→2
- seahash 1→4
- hashbrown 0.14→0.17
- rusqlite 0.29→0.40 (usize→i64 适配)

API 适配:
- build.rs: tonic_build::configure().compile_protos()
- error.rs: bincode::Error → Box<dyn Error>
- logger.rs: usize→i64 (rusqlite 0.40 移除 usize FromSql)

验证: cargo build ✅ cargo test 15/15 ✅ 28项性能测试 ✅

+++ src/master_log_ws.rs
@@ -100,7 +100,7 @@ async fn handle_log_connection(
 struct LogWsResponse {
     status: String,
     message: String,
-    count: usize,
+    count: i64,
 }
 
 /// 处理日志消息
